--- a/Cargo.toml
+++ b/Cargo.toml
@@ -8,7 +8,24 @@ aoc-runner = "0.3.0"
 aoc-runner-derive = "0.3.0"
 
 [dev-dependencies]
-criterion = { version = "2.7.2", package = "codspeed-criterion-compat" }
+codspeed-criterion = { version = "2.7.2", package = "codspeed-criterion-compat" }
+criterion = "0.5.1"
+pprof = { version = "0.13.0", default-features = false, features = [
+    "criterion",
+    "flamegraph",
+    "frame-pointer",
+] }
+
+[profile.release]
+debug = true
+
+[profile.profiling]
+inherits = "release"
+debug = true
+
+[[bin]]
+name = "aoc"
+path = "src/main.rs"
 
 [[bench]]
 name = "day1"