--- a/Cargo.toml
+++ b/Cargo.toml
@@ -6,6 +6,8 @@ edition = "2021"
 [dependencies]
 aoc-runner = "0.3.0"
 aoc-runner-derive = "0.3.0"
+atoi = "2.0.0"
+memchr = "2.7.4"
 
 [dev-dependencies]
 codspeed-criterion = { version = "2.7.2", package = "codspeed-criterion-compat" }
@@ -17,11 +19,21 @@ pprof = { version = "0.13.0", default-features = false, features = [
 ] }
 
 [profile.release]
+codegen-units = 1
+lto = "fat"
+panic = "abort"
+strip = true
+
+[profile.bench]
+inherits = "release"
 debug = true
+strip = false
 
 [profile.profiling]
 inherits = "release"
 debug = true
+strip = false
+
 
 [[bin]]
 name = "aoc"