--- a/Cargo.toml
+++ b/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 aoc-runner = "0.3.0"
 aoc-runner-derive = "0.3.0"
 atoi = "2.0.0"
+fxhash = "0.2.1"
 memchr = "2.7.4"
 
 [dev-dependencies]