--- a/src/day1.rs
+++ b/src/day1.rs
@@ -1,6 +1,9 @@
-use std::collections::HashMap;
-
 use aoc_runner_derive::aoc;
+use atoi::FromRadix10;
+use core::str;
+use memchr::memchr;
+use memchr::memmem;
+use std::collections::HashMap;
 
 // starting with flamegraph
 fn parse_input(input: &str) -> (Vec<u32>, Vec<u32>) {
@@ -9,6 +12,27 @@ fn parse_input(input: &str) -> (Vec<u32>, Vec<u32>) {
         .filter_map(|s| s.split_once("   "))
         .map(|(a, b)| (a.parse::<u32>().unwrap(), b.parse::<u32>().unwrap()))
         .unzip()
+}
+
+fn parse_input_memchr(input: &str) -> (Vec<u32>, Vec<u32>) {
+    let mut data = input[..].as_bytes();
+    let mut a_vec = Vec::with_capacity(1000);
+    let mut b_vec = Vec::with_capacity(1000);
+    let finder = memmem::Finder::new("   ");
+    loop {
+        let Some(separator) = finder.find(data) else {
+            break;
+        };
+        let end = memchr(b'\n', &data[separator..]).unwrap();
+        let a = &data[..separator];
+        let b = &data[separator + 3..separator + end];
+        let a = u32::from_radix_10(a);
+        let b = u32::from_radix_10(b);
+        a_vec.push(a.0);
+        b_vec.push(b.0);
+        data = &data[separator + end + 1..];
+    }
+    (a_vec, b_vec)
 }
 
 fn compute_distance(a_vec: &[u32], b_vec: &[u32]) -> u32 {
@@ -22,7 +46,8 @@ fn compute_distance(a_vec: &[u32], b_vec: &[u32]) -> u32 {
 
 #[aoc(day1, part1, Chars)]
 pub fn part1(input: &str) -> u32 {
-    let (a_vec, b_vec) = parse_input(input);
+    // let (a_vec, b_vec) = parse_input(input);
+    let (a_vec, b_vec) = parse_input_memchr(input);
     compute_distance(&a_vec, &b_vec)
 }
 
@@ -37,7 +62,8 @@ fn compute_similarity_score(a_vec: &[u32], b_vec: &[u32]) -> u32 {
 
 #[aoc(day1, part2)]
 pub fn part2(input: &str) -> u32 {
-    let (a_vec, b_vec) = parse_input(input);
+    // let (a_vec, b_vec) = parse_input(input);
+    let (a_vec, b_vec) = parse_input_memchr(input);
     compute_similarity_score(&a_vec, &b_vec)
 }
 
@@ -50,7 +76,8 @@ mod tests {
 2   5
 1   3
 3   9
-3   3";
+3   3
+";
 
     #[test]
     fn sample1() {