--- a/src/main.rs
+++ b/src/main.rs
@@ -4,7 +4,7 @@
 use aoc_2024::day1;
 
 fn main() -> () {
-    let input = include_str!("../input/2024/full/day1.txt");
+    let input = include_str!("../input/2024/sample/day1.txt");
     let mut out1 = 0;
     let mut out2 = 0;
     for _i in 0..10_000 {