perf: setup profiler to improve perf commit af2e0b8ee1bf489ad0b59ac1d217fca32158dad3 

```
    part1                   time:   [43.465 µs 43.505 µs 43.547 µs]
                            change: [-0.4149% -0.0859% +0.2330%] (p = 0.62 > 0.05)
                            No change in performance detected.
    Found 3 outliers among 100 measurements (3.00%)
      1 (1.00%) high mild
      2 (2.00%) high severe

    part2                   time:   [61.165 µs 61.226 µs 61.290 µs]
                            change: [-0.2693% +0.0569% +0.3443%] (p = 0.73 > 0.05)
                            No change in performance detected.
    Found 5 outliers among 100 measurements (5.00%)
      2 (2.00%) high mild
      3 (3.00%) high severe
```

perf: add memchr and atoi for parsing commit c7852699d6c75fd62ada0a9bfc5c1854479ad318

```
    part1                   time:   [17.639 µs 17.663 µs 17.692 µs]
                            change: [-56.508% -56.300% -56.038%] (p = 0.00 < 0.05)
                            Performance has improved.
    Found 10 outliers among 100 measurements (10.00%)
      4 (4.00%) high mild
      6 (6.00%) high severe

    part2                   time:   [34.221 µs 34.406 µs 34.587 µs]
                            change: [-40.109% -39.854% -39.613%] (p = 0.00 < 0.05)
                            Performance has improved.
    Found 7 outliers among 100 measurements (7.00%)
      7 (7.00%) high mild
```

perf: add fxhash for part 2 commit e2cd1a0dc083c6e67d62d961dadf94fbc03b9a5e
    
```
    part1                   time:   [17.821 µs 17.841 µs 17.861 µs]
                            change: [+0.1907% +0.8146% +1.2743%] (p = 0.00 < 0.05)
                            Change within noise threshold.
    Found 5 outliers among 100 measurements (5.00%)
      2 (2.00%) high mild
      3 (3.00%) high severe

    part2                   time:   [17.138 µs 17.191 µs 17.248 µs]
                            change: [-50.129% -49.925% -49.716%] (p = 0.00 < 0.05)
                            Performance has improved.
    Found 7 outliers among 100 measurements (7.00%)
      2 (2.00%) high mild
      5 (5.00%) high severe
```

perf: FxHashMap::with_capacity_and_hasher commit 271e7cbc8e8063259e65b09afc7a8dd594653147

```
part1                   time:   [17.552 µs 17.569 µs 17.587 µs]
                        change: [-1.5771% -1.1694% -0.7460%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

part2                   time:   [11.568 µs 11.580 µs 11.593 µs]
                        change: [-31.551% -31.312% -31.099%] (p = 0.00 < 0.05)
                        Performance has improved.
```